+++ CHANGELOG.md
@@ -7,12 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `validatePolicy` and `validateAsset` to validation functions
+
 ## [1.0.1] - 2022-10-24
 
 ### Changed
 
 - `getAddressTypeAndPaymentCred` to return undefined instead of `''` when there is not valid `paymentCred`
-
 
 ## [1.0.0] - 2022-09-23